// src/shared/UI/Tooltip/enums.ts
export enum TooltipPosition {
  topStart = 'topStart',
  topCenter = 'topCenter',
  topEnd = 'topEnd',
  bottomStart = 'bottomStart',
  bottomCenter = 'bottomCenter',
  bottomEnd = 'bottomEnd',
  leftStart = 'leftStart',
  leftCenter = 'leftCenter',
  leftEnd = 'leftEnd',
  rightStart = 'rightStart',
  rightCenter = 'rightCenter',
  rightEnd = 'rightEnd',
}